06.14_14:01_Подключена база данных

+++ sso/internal/servises/auth/auth.go
@@ -71,7 +71,7 @@ func (a *Auth) Login(
 	ctx context.Context,
 	email,
 	password string,
-	appID int,
+	appID int64,
 ) (string, error) {
 	const op = "auth.Login"
